RdsNotification moved to RdsNotificationService with backward compatibility
